Extract storage into its own class
