+++ src/js/browser.js
@@ -159,47 +159,6 @@ export var browserUtils = {
     });
   },
 
-  /*
-     * Only chrome has sync, if sync is available use it
-     * otherwise reverts to using local storage
-     * @to-do add a localStorage fallback for browsers that
-     * don't recognize chrome.storage
-     */
-  store: function(key) {
-    return browser.storage.local.set(key);
-  },
-
-  /**
-     * Save an object to local storage via a key
-     *
-     * @param  {string} key    Local storage key
-     * @param  {object} object The object to save
-     * @return {[type]}        [description]
-     */
-  save: function(key, object) {
-    try {
-      browser.storage.local.set(object, function() {
-        var status = document.getElementById("status");
-        status.textContent = key + " saved.";
-        setTimeout(function() {
-          status.textContent = "";
-        }, 750);
-      });
-    } catch (e) {
-      console.log(e);
-    }
-  },
-
-  /**
-     * Get an object from local storage
-     *
-     * @param  {string}   key      The name of key in local storage
-     * @return {Promise}           Promise represents object retrieved from local storage
-     */
-  retrieve: function(key) {
-    return browser.storage.local.get(key);
-  },
-
   /**
      * Convert a path to the fully qualified browser extension URL
      * @param  {string} path - A string to convert into a full URL
